+++ mapentity/views/base.py
@@ -109,7 +109,9 @@ class JSSettings(JSONResponseMixin, TemplateView):
         # Languages
         dictsettings['languages'] = dict(available=dict(app_settings['TRANSLATED_LANGUAGES']),
                                          default=app_settings['LANGUAGE_CODE'])
+        # MAX_CHARACTERS paramters is deprecated : to remove
         dictsettings['maxCharacters'] = app_settings['MAX_CHARACTERS']
+        dictsettings['maxCharactersByField'] = app_settings['MAX_CHARACTERS_BY_FIELD']
         return dictsettings